+++ jlog/logrus-logger.go
@@ -1,0 +1,8 @@
+package jlog
+
+import "github.com/sirupsen/logrus"
+
+type LogrusLogger struct {
+	Log          *logrus.Logger
+	ReportCaller bool
+}